requirements.txt
absl-py==1.3.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
astunparse==1.6.3
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
bokeh==3.0.3
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
colorama==0.4.5
contourpy==1.0.5
cryptography==38.0.4
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
entrypoints==0.4
executing==1.1.1
fastjsonschema==2.16.2
Flask==2.2.2
flatbuffers==22.10.26
fonttools==4.38.0
gast==0.4.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h5py==3.7.0
idna==3.4
idx2numpy==1.2.3
imbalanced-learn==0.9.1
imblearn==0.0
ipykernel==6.16.1
ipympl==0.9.2
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.21.0
jupyter_client==7.4.3
jupyter_core==4.11.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
libsvm-official==3.30.0
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
mpmath==1.2.1
nbclassic==0.4.5
nbclient==0.7.0
nbconvert==7.2.3
nbformat==5.7.0
nest-asyncio==1.5.6
nltk==3.7
notebook==6.5.1
notebook_shim==0.2.0
ntlm-auth==1.5.0
numba==0.56.3
numpy==1.23.4
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
pbr==5.11.0
pickleshare==0.7.5
Pillow==9.2.0
ply==3.11
prometheus-client==0.15.0
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.3
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyrsistent==0.18.1
pyswip==0.2.10
python-dateutil==2.8.2
pytz==2022.5
pywin32==304
pywinpty==2.0.9
PyYAML==6.0
pyzmq==24.0.1
qiskit==0.39.4
qiskit-aer==0.11.2
qiskit-ibmq-provider==0.19.2
qiskit-terra==0.22.3
qtconsole==5.3.2
QtPy==2.2.1
regex==2022.10.31
requests==2.28.1
requests-ntlm==1.1.0
requests-oauthlib==1.3.1
retworkx==0.12.1
rsa==4.9
rustworkx==0.12.1
scikit-learn==1.1.2
scipy==1.9.3
seaborn==0.12.1
Send2Trash==1.8.0
shap==0.41.0
six==1.16.0
sklearn==0.0
slicer==0.0.7
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.5.1
stevedore==4.1.1
sympy==1.11.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.0
terminado==0.17.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
thundersvm @ file:///E:/ML/thundersvm-cu10-0.2.0-py3-none-win_amd64.whl
tinycss2==1.2.1
torch==1.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
tweedledum==1.1.1
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
websockets==10.4
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.14.1
xyzservices==2022.9.0
